move class Client to a separate file
